return ScriptVar_t SVT_CustomWeakRefToStatic references from List<>

+++ src/client/DeprecatedGUI/Menu_Local.cpp
@@ -1346,7 +1346,7 @@ static void updateFeaturesList(CListview* l)
 		updateFeatureListItemColor(item);
 	}
 	if( tLXOptions->customSettings.isSet[FT_Lives] && (int)tLXOptions->customSettings[FT_Lives] < 0 )
-		tLXOptions->customSettings.set(FT_Lives) = WRM_UNLIM;
+		tLXOptions->customSettings.set(FT_Lives) = (int)WRM_UNLIM;
 }
 
 /////////////
